% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\title{Install dependencies}
\usage{
dependencies(scala = F, maven = F)
}
\arguments{
\item{NULL}{}
}
\description{
Installs the dependencies, including HyperSpark source code from
  Github.
}
\examples{
install_dependencies()

}
