% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeHyperSpark.R
\name{executeHyperSpark}
\alias{executeHyperSpark}
\title{Execute HyperSpark}
\usage{
executeHyperSpark()
}
\description{
Executes the framework locally. This requires that packaging has
been successfully executed.
}
\examples{
executeHyperSpark()

}
