% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installDependencies.R
\name{installDependencies}
\alias{installDependencies}
\title{Install dependencies}
\usage{
installDependencies(scala = F, maven = F)
}
\arguments{
\item{NULL}{}
}
\description{
Installs the dependencies, including HyperSpark source code from
  Github.
}
\examples{
install_dependencies()

}
