% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHyperSpark.R
\name{packageHyperSpark}
\alias{packageHyperSpark}
\title{Execute HyperSpark}
\usage{
packageHyperSpark(build_output = F)
}
\arguments{
\item{NULL}{}
}
\description{
Executes HyperSpark.
}
\examples{
package()

}
