% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureHyperspark.R
\name{configureHyperSpark}
\alias{configureHyperSpark}
\title{Configure HyperSpark}
\usage{
configureHyperSpark(
  setProblem,
  data,
  setStoppingCondition,
  stoppingValue,
  setAlgorithms,
  numOfAlgorithms = NONE,
  setNumberOfIterations = NONE,
  setInitialSeeds = NONE,
  setNInitialSeeds = NONE,
  setNDefaultInitialSeeds = NONE,
  setSeedingStrategy = NONE,
  setProperty = NONE,
  setMapReduceHandler = NONE,
  setRandomSeed = NONE,
  setSparkMaster = NONE,
  setAppName = NONE,
  setNumberOfExecutors = NONE,
  setNumberOfResultingRDDPartitions = NONE,
  setDeploymentLocalNoParallelism = NONE,
  setDeploymentLocalNumExecutors = NONE,
  setDeploymentLocalMaxCores = NONE,
  setDeploymentSpark = NONE,
  setDeploymentMesos = NONE,
  setDeploymentYarnClient = NONE,
  setDeploymentYarnCluster = NONE
)
}
\arguments{
\item{setProblem}{Set the problem to be solved.}

\item{data}{Set the data for the application.}

\item{setStoppingCondition}{Set the stopping condition for algorithms to
terminate.}

\item{stoppingValue}{Set the stopping value for the stopping condition.}

\item{setAlgorithms}{Set the algorithms to execute. Always include
parentheses, e.g. GAAlgorithm()}
}
\description{
Lets users configure HyperSpark. The result is a written .scala
file, which is added to the HyperSpark source code.
}
\examples{
hypeRSpark::configureHyperSpark(setProblem = 'PfsProblem',
data = 'inst_ta054.txt', setStoppingCondition = 'TimeExpired',
stoppingValue = 1000, setAlgorithms = 'GAAlgorithm', numOfAlgorithms = 4,
setDeploymentLocalMaxCores = T, setNumberOfIterations = 100,
setNDefaultInitialSeeds = 4)

}
