% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\title{Install dependencies}
\usage{
dependencies()
}
\arguments{
\item{NULL}{}
}
\description{
Downloads and unpacks the framework's source code from GitHub.
This process creates a new folder, called HyperSpark-master in the working
directory.
}
\examples{
dependencies()

}
