% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeHyperSpark.R
\name{executeHyperSpark}
\alias{executeHyperSpark}
\title{Execute HyperSpark}
\usage{
executeHyperSpark()
}
\arguments{
\item{execution_ouput}{If set to TRUE the terminal output of the execution
process is shown in the console, mainly useful for developing and
debugging. The default value is FALSE.}
}
\description{
Executes HyperSpark.
}
\examples{
execute_jar()

}
